* 系统资源获取，动态获取系统资源情况，包括CPU,内存,GPU内存,磁盘IO,GPU利用率,和操作时间 * 系统资源统计可以生成表格，支持多个不同的操作进行系统资源的统计 * 不强制依赖第三方库，如 SPDLOG 和 NVML,如果没有也不至于报错

+++ src/file_tools.cpp
@@ -10,6 +10,7 @@
 #include "include/file_tools.h"
 
 #include <fstream>
+#include <iostream>
 using namespace jade;
 bool FileTools::CreateDirectory(const std::string& path)
 {
@@ -106,7 +107,7 @@ bool FileTools::writeBinaryToFile(const string& path, const float* data, const i
 {
     // 写入二进制文件
     std::ofstream outFile(path, ios::out | ios::binary);
-    if (!outFile.write(reinterpret_cast<const char*>(data), size * sizeof(float))) {
+    if (!outFile.write(reinterpret_cast<const char*>(data), static_cast<long long>(size* sizeof(float)))) {
         std::cerr << "文件路径为:" << path << "写入文件失败！" << std::endl;
         return false;
     }
